move BinaryCheck into check package to fix import cycle
