+++ internal/check/binary.go
@@ -1,11 +1,9 @@
-package checks
+package check
 
 import (
 	"context"
 	"fmt"
 	"os/exec"
-
-	"github.com/vidya381/devcheck/internal/check"
 )
 
 type BinaryCheck struct {
@@ -16,19 +14,19 @@ func (c *BinaryCheck) Name() string {
 	return fmt.Sprintf("%s installed", c.Binary)
 }
 
-func (c *BinaryCheck) Run(_ context.Context) check.Result {
+func (c *BinaryCheck) Run(_ context.Context) Result {
 	_, err := exec.LookPath(c.Binary)
 	if err != nil {
-		return check.Result{
+		return Result{
 			Name:    c.Name(),
-			Status:  check.StatusFail,
+			Status:  StatusFail,
 			Message: fmt.Sprintf("%s not found on PATH", c.Binary),
 			Fix:     fmt.Sprintf("Install %s and make sure it is on your PATH", c.Binary),
 		}
 	}
-	return check.Result{
+	return Result{
 		Name:    c.Name(),
-		Status:  check.StatusPass,
+		Status:  StatusPass,
 		Message: fmt.Sprintf("%s is installed", c.Binary),
 	}
 }
