+++ internal/check/binary_test.go
@@ -1,16 +1,14 @@
-package checks
+package check
 
 import (
 	"context"
 	"testing"
-
-	"github.com/vidya381/devcheck/internal/check"
 )
 
 func TestBinaryCheck_Pass(t *testing.T) {
 	c := &BinaryCheck{Binary: "go"}
 	result := c.Run(context.Background())
-	if result.Status != check.StatusPass {
+	if result.Status != StatusPass {
 		t.Errorf("expected pass, got %v: %s", result.Status, result.Message)
 	}
 }
@@ -18,7 +16,7 @@ func TestBinaryCheck_Pass(t *testing.T) {
 func TestBinaryCheck_Fail(t *testing.T) {
 	c := &BinaryCheck{Binary: "definitelynotabinary12345"}
 	result := c.Run(context.Background())
-	if result.Status != check.StatusFail {
+	if result.Status != StatusFail {
 		t.Errorf("expected fail, got %v", result.Status)
 	}
 }
